Move some helper logic out of update-chip8-archive.js
